+++ R/run.R
@@ -129,6 +129,26 @@ run_internal_native <- function(cmd,
                                 env_name = "condathis-env",
                                 verbose = TRUE,
                                 stdout = "|") {
+
+  if (isTRUE(base::Sys.info()["sysname"] == "Windows")) {
+    micromamba_bat_path <- fs::path(get_install_dir(), "condabin", "micromamba", ext = "bat")
+    if (isFALSE(fs::file_exists(micromamba_bat_path))) {
+      catch_res <- rlang::catch_cnd(
+        expr = {
+          native_cmd(
+            conda_cmd = "run",
+            conda_args = c("-n", "condathis-env"),
+            cmd = "dir", verbose = FALSE, stdout = NULL
+          )
+        }
+      )
+      mamba_bat_path <- fs::path(get_install_dir(), "condabin", "mamba", ext = "bat")
+      if (isTRUE(fs::file_exists(mamba_bat_path)) &&
+          isFALSE(fs::file_exists(micromamba_bat_path))) {
+        fs::file_copy(mamba_bat_path, mamba_bat_path, overwrite = TRUE)
+      }
+    }
+  }
   px_res <- native_cmd(
     conda_cmd = "run",
     conda_args = c(